requirements.txt
fastapi
uvicorn
python-pptx
httpx
python-multipart
google-generativeai